fix: Raise exception instead of error log when sniffer did not reply properly to setup commands
When sniffer did not reply properly to setup commands, it's useless. Thus, is better
to inform finish work insted of keeping ti running.
As the exception is in serial_reader thread, additional method is_running has been implemented
to give user ability to confirm that sniffer is up and running after startup.

Signed-off-by: Tomasz Tyzenhauz <tomasz.tyzenhauz@nordicsemi.no>